move old reading code to rfi-mitigation
