+++ champss/sps-common/pyproject.toml
@@ -11,7 +11,6 @@ exclude = [
 
 [tool.poetry.dependencies]
 python = ">=3.10,<3.12"
-spshuff = {path = "../spshuff", develop=true}
 astropy = "~6.1"
 attrs = "^22.2.0"
 h5py = "~3.9"
